Enhance the ".timer" command in the CLI to accept the "once" argument, and so that it leaves its last real-time result in the $TIMER variable.  Also fix a harmless warning from an earlier check-in.
FossilOrigin-Name: 8ad7dffc261cbb3f7c9b4d38b04282c805c1c047fb431076f85e577c81f21574

+++ src/select.c
@@ -3497,7 +3497,7 @@ static int multiSelectByMerge(
       assert( pItem->u.x.iOrderByCol>0 );
       assert( pItem->u.x.iOrderByCol<=p->pEList->nExpr );
       aPermute[i] = pItem->u.x.iOrderByCol - 1;
-      if( aPermute[i]!=i-1 ) bKeep = 1;
+      if( aPermute[i]!=(u32)i-1 ) bKeep = 1;
     }
      if( bKeep==0 ){
       sqlite3DbFreeNN(db, aPermute);
